+++ src/Loading/index.js
@@ -1,5 +1,4 @@
-import React from 'react'
-import styles from './styles.css'
+import styles from './styles.module.css'
 
 export default () => (
   <div className={styles.spinner}>
